requirements.txt
asgiref==3.2.10
Django==3.0.8
pytz==2020.1
sqlparse==0.3.1
Pillow==7.2.0

django-filebrowser-no-grappelli==3.8.0
django-tinymce4-lite==1.7.5